除smiley, httpClient, luascirpt-module及demo外，其它模块在mingw下静态库模式编译通过。 删除qmake，以后主要使用cmake。 修复在使用mingw编译时发现的几个小BUG。 ver:4.5.4.0

+++ components/render-gdi/render-gdi.cpp
@@ -957,7 +957,7 @@ namespace SOUI
 				RECT rc2 = {x-sz.cy, y, x, y+sz.cx};
 				rc = rc2;
 			}else{
-				RECT rc2 = {0,0, m_curBmp->Width(),m_curBmp->Height()};
+				RECT rc2 = {0,0, (long)m_curBmp->Width(),(long)m_curBmp->Height()};
 				rc = rc2;
 			}
 			DCBuffer dcBuf(m_hdc,&rc,m_curColor.a);
